Bunch of changes to get the trivial object space working again.
The deepest: our builtins know nothing of old style classes.
Unfortunately, the trivial object space exported the Exceptions from
the underlying interpreter, which were old style classes.

So, like the std object space, clone the interpreters exception
heirachy as new-style classes.  This means we need to convert the
exceptions raised by the wrapped operations into our exceptions --
this is the point of the 'reraise' function.

Also, add __get__ to nufun, add a few object space methods that have
appeared recently and a check that we get them all and a few cosmetic
bits.

With these changes, the only failure in the test suite is due to 2.2.X
not supporting extended slices.