Add a dir() function to builtins.
This is modeled directly after CPython's dir function.

It is currently incomplete. dir() with no args doesn't work
yet, and trying to pull dir() of an object exposes some
problem with the builtin isinstance() method.

I don't have time to track down the errors at this point,
but I'm committing it anyway, as it is unlikely to break anything
which is currently working.