fixed some bugs.
It seems that the python implementation wont add new errors to the python regression tests.

If someone would be so kind as to run the tests for me (just rename inprogress__codecs.py to _codecs.py), I will try to implement Armin's suggested solution for not modifying the encodings module.

Test_unicode.py should also work (a part from two failures which should be fixed in unicodetype.py)