Conversions between an instance of a class and an instance of a subclass and back.  test_virutal_methods() passes.
Simplification of the varmapping mess, which was provoking strange errors. The
problem is that two variables may have different SomeXxx annotations but the
same low-level type; in this case we cannot really rename one into the other,
because it confuses the rtyper -- which might find the wrong annotation and
representation for the variable.  (It also confuses people staring at flow
graphs quite a lot.)