a few enhancements which I allowed myself to do on the ferry and in Kiel at my parent's  (blush)
- removed all explicit typing, because this is the beginning of all evil!
- upgraded all implementations to the latest C source in CVS.
- implemented Karatsuba multiplication
- added optimizations for a*a
- added optimized pow()
- added more comments
- a lot of new tests

this appears to be almost State of Python Art.

XXX still to do at some time: I think we should
re-adopt CPython's use of a signed length to store
sign and size of the digit array. It is more efficient
to carry explicit length than to really have to adjust
a list. Also I think the list will get optimized
away at some point, and we want an explicit length, anyway.
But this isn't urgent.