(hpk,nik)
only run regression tests in verbose mode if
we are not running an output tests (mirrors
cpython regrtest.py behaviour)

disabled test_zipimport.py because it depends
on zlib which is a non-core module.

improved the way tests are handled when
invoking '-E' but there still seems to
be something severly broken.  tests
with -E pass while they fail with
a plain run.