overallocation of low-level lists done quite right I guess. There is still the possibility to gain a few cycles if we introduce a low-level move function that shifts references around without counting; this nevertheless doesn't apply to the boehm case.
On Windows, using Boehm, this gets us an pystone acceleration of

           576 to 671 pystones

Most probably because we are using pus/pop on the frame stack,so this
will probably vanish again when we do stacks the right way.

Going to remove over-allocation from listobject.py, doubling this
makes no sense at all.