I think that this is doing what I described in the pypy-dev mail, but I'm not sure.  The factorial test of test_jit_tl.py still becomes 'return 5040' instead of staying a loop...
Before computing each block, we compute a 'key' which is derived from the
current state's fixed constants.  Instead of only one residual block per
state, there is one residual block per 'key'.  The residual block in question
has inputargs that are constants -- at least for each fixed constant, but
possibly for more, if policy.const_propagate is True.

When we consider a link that should go to a given block, we compute the 'key'
and check if there is already a corresponding residual block; if so, we check
the constants that have been put in the inputargs.  If they don't match the
new link's constants, we throw away the existing residual block and compute a
new one with less constants in its inputargs.

These recomputations are based on the official 'key', so that links with
different *fixed* constants don't interfere with each other.