(arigo, cfbolz):
Started to work on the FrameworkGCTransformer.  It inserts matching
gc_push_roots/gc_pop_roots operations around calls and mallocs.

We had to extend the 'cleanup' model (again :-/); now an operation's cleanup is
a tuple (finally-operations, except-operations), where the former are always
run after the operation even if it raises, and the latter are run *only* if it
raises.

The Pygame viewer now displays these operations too.