so.  make a test using a static root pass.
the easy part, it turned out was finding the static roots and making sure they
were always part of the root set.

the harder part was realizing that there was a bug in
FrameworkGcPolicy.common_gcheader_initdata -- in the case of an embedded
structure you don't want the typeid to be that of the embedded structure, you
want it to be that of the structure doint the embedding.

the most aggravating part of all this was that i wasn't finding the types of
all the constants present in the graph, so I work harder.  i'm not sure i've
found them all yet, but the test passes at least.

this would all probably have gone quicker if cfbolz had been in #pypy today --
i've basically written a stripped down version of
pypy.rpython.memory.convertlltype.  ho hum.