pypy-c _finally_ built, so add some (not too good) performance numbers.
looking at the c source reveals that the inlining is not as thorough as we
might have liked for a sort of "haha" reason: direct_calls in 'finally'
cleanups are not inlined (presumably 'except' cleanups have the same problem).

maybe it should be an except block and a cleanup block, or something.