Code I have been working on for to long without staying in contact with the other PyPy developers.
It's purpose is to use C++ libraries with (r)ctypes.
The code will generate .h, .cpp and .py files out of a class specification.

This specs look similar to boost.Python . From my experience with prior
project I know that C++ header files with all it's inline function, templates,
etc. quickly become to complicated to handle automaticly. Even with a tool such
as gccxml there are often no clear cut solutions. That is the why.

As an example you can look at wrap_llvm.py , which when executed creates libllvm.h
libllvm.cpp (the C API) and llvm.py (the Python classes which mimic the C++ API's
behaviour)

A pointer to the original C++ object is stored in the 'instance' attribute of the
Python class. Whenever one of these Python objects is used as a parameter ctypes
automatically uses the objects 'instance' attribute instead. (see wrapper.py from_param)
I am no sure if we (can) support this in rctypes as well, otherwise this tool could
do it for us.

I hope this code can be of use and that I did not isolate me too much for too long!