Refactored the extregistry interface.  See mail to pypy-dev.
This is a massive diff, mostly because all rctypes/a*.py are
entierely modified.  Tedious, but the result looks quite nicer
than before in my humble opinion.

The original motivation was a problem with annotating built-in
functions: the 'analyser' attribute cannot be a local function
with a closure, at least as used in rctypes, because
re-building another closure with exactly the same values in
the closure makes a non-equal function object.  Instead, in
the new extregistry world, this 'analyser' is a bound method
on an ExtRegistryEntry instance.

Tests showing this problem:
  * rctypes.test.test_rfunc.test_annotate_reflow_bug
  * targetdemomodule.py with the change in the cpyobjspace

I then remembered the hard way that CPython bound method instances don't
have a sane __eq__ either, so I bite the bullet and wrote a custom
InstanceMethod class to use as the 'analyser'...