issue184 testing
Build a pypy-c that should be movable to a new location or a new machine.
It looks for its library paths (lib-python and pypy/lib) by walking dirs
upwards until it finds them.  It also looks if these paths can be found
in a subdirectory "share/pypy-0.9" along the way.