issue300 testing
as experience has shown that it is much easier for a backend to provide
int_floordiv that truncates towards 0, define int_floordiv so and make int /
int do the cpython thing by means of scary code in rint.py.

add a small special case for when both operands are known to be non-negative.

this also subsumes issue297.

the change to the squeak backend is very largely a guess, that to cl rather
less so, but neither have been tested.