issue 300 testing
the root problem was that the '_zer' variants of the division and modulo
operations weren't getting their results adjusted, so be cleaner and more
comprehensive there.  the fix was held up for a while finding out that the
llinterp's implementation of op_int_floordiv_ovf and so on didn't get updated
when they should have (in rev 40556), but that's fixed now.  add (hopefully)
comprehensive tests.