For now, reduce the write_barrier() interface. I don't know if I'm making some GCs impossible to write, but at least it fixes a problematic crash -- and I could kill a page of obscure code, probably not too bad :-)
The issue is that if the write barrier has to write the new
field value itself via the address, we get strange C code
manipulating "void* + offset".  The C compiler's strict
aliasing rules don't realize that it's a write to a field
of a structure and may reorder a subsequent read before
the write...  see gcc option -fno-strict-aliasing.
With this change, the write barrier is no longer responsible
for actually writing the field - we still emit a bare_setfield
which the C compiler understands correctly.