Fix select readability of write pipes
Add a test asserting that the write end of a pipe is indicated
as readable by select when the read end has been closed.  The
read set indicates either that data may be available or that
the file descriptor is at EOF.

There are probably still some bugs in select.select() (particularly the
exception set).