Merge the 'logging2' branch.  Add a general way to record logging events to pypy.rlib.debug.  In addition to debug_print() there is now debug_start() and debug_stop() to mark the start/stop of a category.
    svn merge -r68854:68901 svn+ssh://codespeak.net/svn/pypy/branch/logging2

Use it in the GC (replaces the gcconfig.debugprint config option) and in
the JIT (similar to jitprof.py, which is not dead yet but obsoleting).

To enable logging run the compiled program with 'PYPYLOG=:file'.  To
enable lightweight logging only, suitable for profiling, run it with
'PYPYLOG=file'.  See translator/c/src/debug.h for more options.

To turn the profiling logs into nice graphics, see pypy/tool/logparser.py.