Merge branch msvc-asmgcroot: support asmgcroot for the Microsoft compiler This is necessary for a framework gc to work with the jit. (the alternatives being --gc=boehm, or --cc=mingw32)
Now "translate.py -Ojit" should work out of the box.

The MASM assembler (ml.exe) is used. It may come with your installation of Visual Studio
or with the Platform SDK; otherwise it can be installed from http://www.masm32.com/

The script trackgcroot.py was refactored, hopefully in a more object-oriented way,
mostly because the assembler generated by the MSVC is very different from gcc.
Also, the markers inserted in the code to indicater gc roots are different, because
the compiler seems to optimize even inline __asm blocks.

Moreover an "extern" function pointer often does not point to the function code,
but inside a relocation map which contains a jmp to the real code.