Fix one of the issues that I've been fighting for since a week: the accesses to the 'forced' field of virtualrefs is not recorded in the effectinfo, thus from time to time invalid code is generated -- by delaying the SETFIELD_GC for too long.  This is how force_virtual() occasionally showed invalid VIRTUAL_REF structures: because the structure was not fully updated yet.
This fix also simplifies things: the effectinfos that used to have the
flag 'forces_virtual_or_virtualizable' are now always stored just as
None.  I think that there is not a lot of benefits from optimizing field
reads or writes around calls to functions that can force anyway.

This is a merge from r71373+r71374 in branch/debug-vref2.