Merge branch/debug-vref3:
Revert r71374, as it appears that there are cases in which
the performance matters -- i.e. the fact that some fields
may remain virtual across a call to a function that can
force virtualrefs or virtualizables.  (According to fijal
he sees a 30% difference on one benchmark.)

Fix it differently in optimizeopt.py, by saying that calls
to such functions always force at least the 'forced' field
of virtualrefs (but not necessarily other fields).