Merge branch/guard-value-counting-2.
For GUARD_VALUE failures, gives one counter per actual value instead of
just a single counter.  This should focus compilation on the most
commonly seen values, and prevent compilation on the rarely seen values.
As an extreme example, the new test in test_basic checks that if the
GUARD_VALUE sees always different values, the failure never gets
compiled.  The same effect in PyPy would be that if a class' VersionTag
changes all the time, then we should not compile, instead of repeatedly
compiling versions of the code that are never reused.