Merge branch/stackovf: introduce the StackOverflow exception.
It is a subclass of RuntimeError, raised in RPython programs for stack
overflows, and defined in pypy.rlib.rstackovf.

There is a bit of magic in pypy.rlib.rstackovf to let you say "except
StackOverflow:" and still have it caught when running untranslated (in which
case you get of course just CPython's RuntimeError).  But it allows some
cleanups a bit here and there, and "pypy-c" now can confidently print
"RuntimeError: maximum recursion depth exceeded" instead of "RuntimeError:
internal error (stack overflow?)".