Merge the branch/kill-caninline, whose purpose is to kill the "can_inline" hack in the JIT.  This makes some benchmarks faster because a Python function can be inlined into another one even if it contains a loop, and inlining stops only when (and if) the inner loop is reached.
Left open: there is still the code to sometimes start JIT tracing from
the start of a Python function; this is now only useful in case a
previous tracing gave up because the trace was too long.  Should ideally
be handled differently.

Also, JITting Python generators is fully disabled for now.