issue543 resolved
Kill the pure Python _sre.py.  It may work or not work, but it has no
test.  Moreover it's based on the "old" algorithm, now gone.  If you
want a pure Python replacement for _sre, there is a rough one in
pypy/rlib/rsre/rsre_re.py, used by tests only for now.