Allow the creation of io.xxxBase objects.
CPython can inherit tp_new from the base class, because
it assumes that zeroing the memory is a sane default for most types.
The closest thing we could do in PyPy is to call the __init__ function
with no argument, but it's difficult since the typedef has no reference
to the W_Class. Meanwhile,
    __new__ = GenericNew(W_Class)
is not too bad. If it works, I'll move it to typedef.py.

--HG--
branch : fast-forward